third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/contentSecurityPolicy/multiple-policies-with-nonce-expected.txt

CONSOLE ERROR: The Content Security Policy 'script-src 'self'' was delivered in report-only mode, but does not specify a 'report-uri'; the policy will have no effect. Please either add a 'report-uri' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE ERROR: line 4: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'nonce-abcd1234'". Either the 'unsafe-inline' keyword, a hash ('sha256-LzK6Wilfo0IyS30ovav5qfVnk5gyXUKW3gX9Kj9aEig='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 4: [Report Only] Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self'". Either the 'unsafe-inline' keyword, a hash ('sha256-LzK6Wilfo0IyS30ovav5qfVnk5gyXUKW3gX9Kj9aEig='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 7: [Report Only] Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self'". Either the 'unsafe-inline' keyword, a hash ('sha256-bQkhZIU40UiVfL5eqBARanGzrKeiIBRVwl4k+weB/v4='), or a nonce ('nonce-...') is required to enable inline execution.

This is a testharness.js-based test.
PASS This script block has a matching nonce, and should execute. 
Harness: the test ran to completion.
